完善京东

+++ app/src/main/java/com/bw/my_jingdong/mvp/home/model/bean/ProductDetailsBean.java
@@ -5,9 +5,9 @@ public class ProductDetailsBean {
 
     /**
      * msg :
-     * seller : {"description":"我是商家1","icon":"http://120.27.23.105/images/icon.png","name":"商家1","productNums":999,"score":4.9,"sellerid":1}
+     * seller : {"description":"我是商家17","icon":"http://120.27.23.105/images/icon.png","name":"商家17","productNums":999,"score":5,"sellerid":17}
      * code : 0
-     * data : {"bargainPrice":11800,"createtime":"2017-10-10T17:33:37","detailUrl":"https://item.m.jd.com/product/4338107.html?utm#_source=androidapp&utm#_medium=appshare&utm#_campaign=t#_335139774&utm#_term=QQfriends","images":"https://m.360buyimg.com/n0/jfs/t6700/155/2098998076/156185/6cf95035/595dd5a5Nc3a7dab5.jpg!q70.jpg","itemtype":0,"pid":57,"price":5199,"pscid":40,"salenum":4343,"sellerid":1,"subhead":"【i5 MX150 2G显存】全高清窄边框 8G内存 256固态硬盘 支持指纹识别 预装WIN10系统","title":"小米(MI)Air 13.3英寸全金属轻薄笔记本(i5-7200U 8G 256G PCle SSD MX150 2G独显 FHD 指纹识别 Win10）银\r\n"}
+     * data : {"bargainPrice":111.99,"createtime":"2017-10-14T21:39:05","detailUrl":"https://item.m.jd.com/product/4719303.html?utm_source=androidapp&utm_medium=appshare&utm_campaign=t_335139774&utm_term=QQfriends","images":"https://m.360buyimg.com/n0/jfs/t9004/210/1160833155/647627/ad6be059/59b4f4e1N9a2b1532.jpg!q70.jpg|https://m.360buyimg.com/n0/jfs/t7504/338/63721388/491286/f5957f53/598e95f1N7f2adb87.jpg!q70.jpg|https://m.360buyimg.com/n0/jfs/t7441/10/64242474/419246/adb30a7d/598e95fbNd989ba0a.jpg!q70.jpg","itemtype":1,"pid":1,"price":118,"pscid":1,"salenum":0,"sellerid":17,"subhead":"每个中秋都不能简单，无论身在何处，你总需要一块饼让生活更圆满，京东月饼让爱更圆满京东自营，闪电配送，更多惊喜，快用手指戳一下","title":"北京稻香村 稻香村中秋节月饼 老北京月饼礼盒655g"}
      */
 
     private String msg;
@@ -49,19 +49,19 @@ public class ProductDetailsBean {
 
     public static class SellerBean {
         /**
-         * description : 我是商家1
+         * description : 我是商家17
          * icon : http://120.27.23.105/images/icon.png
-         * name : 商家1
+         * name : 商家17
          * productNums : 999
-         * score : 4.9
-         * sellerid : 1
+         * score : 5
+         * sellerid : 17
          */
 
         private String description;
         private String icon;
         private String name;
         private int productNums;
-        private double score;
+        private int score;
         private int sellerid;
 
         public String getDescription() {
@@ -96,11 +96,11 @@ public class ProductDetailsBean {
             this.productNums = productNums;
         }
 
-        public double getScore() {
+        public int getScore() {
             return score;
         }
 
-        public void setScore(double score) {
+        public void setScore(int score) {
             this.score = score;
         }
 
@@ -115,39 +115,38 @@ public class ProductDetailsBean {
 
     public static class DataBean {
         /**
-         * bargainPrice : 11800
-         * createtime : 2017-10-10T17:33:37
-         * detailUrl : https://item.m.jd.com/product/4338107.html?utm#_source=androidapp&utm#_medium=appshare&utm#_campaign=t#_335139774&utm#_term=QQfriends
-         * images : https://m.360buyimg.com/n0/jfs/t6700/155/2098998076/156185/6cf95035/595dd5a5Nc3a7dab5.jpg!q70.jpg
-         * itemtype : 0
-         * pid : 57
-         * price : 5199
-         * pscid : 40
-         * salenum : 4343
-         * sellerid : 1
-         * subhead : 【i5 MX150 2G显存】全高清窄边框 8G内存 256固态硬盘 支持指纹识别 预装WIN10系统
-         * title : 小米(MI)Air 13.3英寸全金属轻薄笔记本(i5-7200U 8G 256G PCle SSD MX150 2G独显 FHD 指纹识别 Win10）银
-
+         * bargainPrice : 111.99
+         * createtime : 2017-10-14T21:39:05
+         * detailUrl : https://item.m.jd.com/product/4719303.html?utm_source=androidapp&utm_medium=appshare&utm_campaign=t_335139774&utm_term=QQfriends
+         * images : https://m.360buyimg.com/n0/jfs/t9004/210/1160833155/647627/ad6be059/59b4f4e1N9a2b1532.jpg!q70.jpg|https://m.360buyimg.com/n0/jfs/t7504/338/63721388/491286/f5957f53/598e95f1N7f2adb87.jpg!q70.jpg|https://m.360buyimg.com/n0/jfs/t7441/10/64242474/419246/adb30a7d/598e95fbNd989ba0a.jpg!q70.jpg
+         * itemtype : 1
+         * pid : 1
+         * price : 118
+         * pscid : 1
+         * salenum : 0
+         * sellerid : 17
+         * subhead : 每个中秋都不能简单，无论身在何处，你总需要一块饼让生活更圆满，京东月饼让爱更圆满京东自营，闪电配送，更多惊喜，快用手指戳一下
+         * title : 北京稻香村 稻香村中秋节月饼 老北京月饼礼盒655g
          */
 
-        private double bargainPrice;
+        private float bargainPrice;
         private String createtime;
         private String detailUrl;
         private String images;
         private int itemtype;
         private int pid;
-        private double price;
+        private int price;
         private int pscid;
         private int salenum;
         private int sellerid;
         private String subhead;
         private String title;
 
-        public double getBargainPrice() {
+        public float getBargainPrice() {
             return bargainPrice;
         }
 
-        public void setBargainPrice(double bargainPrice) {
+        public void setBargainPrice(float bargainPrice) {
             this.bargainPrice = bargainPrice;
         }
 
@@ -191,11 +190,11 @@ public class ProductDetailsBean {
             this.pid = pid;
         }
 
-        public double getPrice() {
+        public int getPrice() {
             return price;
         }
 
-        public void setPrice(double price) {
+        public void setPrice(int price) {
             this.price = price;
         }
